refactor openweather helper
